backend/requirements.txt
ultralytics
roboflow
fastapi
uvicorn
python-multipart
requests
python-dotenv
pillow
